requirements.txt
django~=1.11
django-ckeditor
requests==2.20
